refactor: rewrite runtime-core in TypeScript with intermediate build step
