+++ packages/runtime-core/config/api-extractor.json
@@ -1,6 +1,6 @@
 {
   "$schema": "https://developer.microsoft.com/json-schemas/api-extractor/v7/api-extractor.schema.json",
-  "mainEntryPointFilePath": "<projectFolder>/src/index.d.ts",
+  "mainEntryPointFilePath": "<projectFolder>/lib/index.d.ts",
   "bundledPackages": ["@douglasneuroinformatics/libui-form-types"],
   "apiReport": {
     "enabled": false
