--- packages/runtime-core/src/define.js
@@ -1,5 +1,0 @@
-export function defineInstrument(def) {
-  return Object.assign(def, {
-    __runtimeVersion: 1
-  });
-}
