--- packages/runtime-core/src/index.d.ts
@@ -1,9 +1,0 @@
-export type * from './define.d.ts';
-export type * from './i18n.d.ts';
-export type * from './types/core.d.ts';
-export type * from './types/instrument.base.d.ts';
-export type * from './types/instrument.core.d.ts';
-export type * from './types/instrument.form.d.ts';
-export type * from './types/instrument.interactive.d.ts';
-export type * from './types/instrument.series.d.ts';
-export type * from './utils.d.ts';
