--- packages/runtime-core/src/index.js
@@ -1,3 +1,0 @@
-export * from './define.js';
-export * from './i18n.js';
-export * from './utils.js';
