+++ packages/runtime-core/src/types/instrument.series.ts
@@ -1,7 +1,7 @@
 import type { Merge } from 'type-fest';
 
-import type { Language } from './core.d.ts';
-import type { BaseInstrument, InstrumentLanguage, ScalarInstrumentInternal } from './instrument.base.d.ts';
+import type { Language } from './core.js';
+import type { BaseInstrument, InstrumentLanguage, ScalarInstrumentInternal } from './instrument.base.js';
 
 /** @beta */
 declare type SeriesInstrument<TLanguage extends InstrumentLanguage = InstrumentLanguage> = Merge<
