--- packages/runtime-core/src/utils.d.ts
@@ -1,4 +1,0 @@
-import type { SnakeCasedProperties } from 'type-fest';
-
-/** @alpha */
-export declare function asSnakeCase<T extends { [key: string]: any }>(target: T): SnakeCasedProperties<T>;
