+++ packages/runtime-core/tsconfig.json
@@ -1,8 +1,10 @@
 {
   "extends": "../../tsconfig.base.json",
   "compilerOptions": {
-    "checkJs": false,
-    "lib": ["ESNext", "DOM", "DOM.Iterable"]
+    "incremental": false,
+    "lib": ["ESNext", "DOM", "DOM.Iterable"],
+    "module": "Node16",
+    "moduleResolution": "Node16"
   },
   "include": ["src/**/*"]
 }
